DAO/backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
frozendict==2.3.8
moralis==0.1.35
python-dateutil==2.7.5
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.3.0
urllib3==1.26.15
